requirements_tf.txt
keras==2.15.0
matplotlib==3.8.3
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==9.5.0
pyIFD==0.0.3
rawpy==0.19.1
scikit_learn==1.4.1.post1
scipy==1.12.0
tensorflow==2.15.0.post1
